# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "bilemo-api",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://bilemo.fr/api",
            "description": "Api BileMo"
        }
    ],
    "paths": {
        "/products": {
            "get": {
                "operationId": "App\\Controller\\ProductController::products",
                "responses": {
                    "200": {
                        "description": "Liste des produits",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "La ressource n'existe pas"
                    },
                    "401": {
                        "description": "Jeton authentifié échoué / invalide"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/products/{id}": {
            "get": {
                "operationId": "App\\Controller\\ProductController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la ressource",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Un produit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "La ressource n'existe pas"
                    },
                    "401": {
                        "description": "Jeton authentifié échoué / invalide"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/users": {
            "get": {
                "operationId": "App\\Controller\\UserController::users",
                "responses": {
                    "200": {
                        "description": "Liste des utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "La ressource n'existe pas"
                    },
                    "401": {
                        "description": "Jeton authentifié échoué / invalide"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            },
            "post": {
                "operationId": "App\\Controller\\UserController::add",
                "responses": {
                    "201": {
                        "description": "Creation d'un utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "La ressource n'existe pas"
                    },
                    "401": {
                        "description": "Jeton authentifié échoué / invalide"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            },
            "delete": {
                "operationId": "App\\Controller\\UserController::delete",
                "responses": {
                    "200": {
                        "description": "Suppression d'un utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "La ressource n'existe pas"
                    },
                    "401": {
                        "description": "Jeton authentifié échoué / invalide"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "operationId": "App\\Controller\\UserController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la ressource",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Un produit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "La ressource n'existe pas"
                    },
                    "401": {
                        "description": "Jeton authentifié échoué / invalide"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "price": {
                        "type": "string",
                        "nullable": true
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "clients": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "client": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "name": "bearer",
                "in": "header",
                "bearerFormat": "JWT"
            }
        }
    }
}